refactor: Reorganizar componentes
