+++ frontend/src/main.ts
@@ -2,10 +2,11 @@ import CheckClose from '@/components/CheckClose.vue';
 import detectarPosicaoCongelada from '@/diretivas/detectarPosicaoCongelada';
 import ErrorComponent from '@/components/ErrorComponent.vue';
 import FormErrorsList from '@/components/FormErrorsList.vue';
-import LabelFromYup from '@/components/LabelFromYup.vue';
+import LabelFromYup from '@/components/camposDeFormulario/LabelFromYup.vue';
 import LoadingComponent from '@/components/LoadingComponent.vue';
 import MigalhasDePão from '@/components/MigalhasDePao.vue';
 import SmaeLink from '@/components/SmaeLink.vue';
+import SmaeText from '@/components/camposDeFormulario/SmaeText.vue';
 import TítuloDePágina from '@/components/TituloDaPagina.vue';
 import selecionarMultiplasOpcoes from '@/diretivas/selecionarMultiplasOpcoes';
 import type { RequestS } from '@/helpers/requestS';
@@ -153,6 +154,7 @@ app.component('LoadingComponent', LoadingComponent);
 app.component('MigalhasDePão', MigalhasDePão);
 app.component('MigalhasDePao', MigalhasDePão);
 app.component('SmaeLink', SmaeLink);
+app.component('SmaeText', SmaeText);
 app.component('TítuloDePágina', TítuloDePágina);
 app.component('TituloDePagina', TítuloDePágina);
