+++ frontend/src/views/comunicadosGerais/partials/ComunicadosGeraisFiltros.vue
@@ -3,8 +3,6 @@ import { computed, watch } from 'vue';
 import { format, subDays } from 'date-fns';
 import { useRoute, useRouter } from 'vue-router';
 import { Field, useForm, ErrorMessage } from 'vee-validate';
-
-import LabelFromYup from '@/components/LabelFromYup.vue';
 import FormularioQueryString from '@/components/FormularioQueryString.vue';
 import { comunicadosGeraisFiltrosSchema, comunicadosGeraisFiltrosSchemaTipoOpcoes } from '@/consts/formSchemas';
